Make the FallbackBitmapLoader the default
