+++ Splat/Splat-NetCore45.csproj
@@ -106,7 +106,6 @@
     <Compile Include="PlatformModeDetector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WinRT\Bitmaps.cs" />
-    <Compile Include="WinRT\FallbackBitmapLoader.cs" />
     <Compile Include="WinRT\Color.cs" />
     <Compile Include="WinRT\Point.cs" />
     <Compile Include="WinRT\Rect.cs" />
